MemRW: "call_inject": removed KEDR_CHECK_ARG(), it is unreliable
I have observed on some systems that the compiler does sometimes insert
instruction that modify %eax/%rdi before the comparison in this macro. This
makes no harm because the value of the function's argument is still handled
properly but WARN_ON_ONCE() triggers here.

It seems, the implementation of KProbes does not checks it the arguments
have been passed correctly in the registers. Let's hope its authors knew
what they were doing.