Coverage tool: Interpret a coverage trace in a way which doesn't remember 'TN:' division, only 'SN:' division.
Some codebases (like Linux kernel) include the same source(!) file into two compilation units in a way,
when it is impossible to distinguish this units according to the coverage trace.
In such cases it is impossible to generate unique "name" for 'TN:', so add/diff/new-coverage are unable to operate on
these compilation units separately.

Since final statistics is extracted from the "merged" information about compilation units, it could be easier to read the trace
in that merged format.

This is intermediate commit: new files are not used yet.